Experimental solution to ray/jax failure
